requirements.txt
langchain-openai
langchain-chroma
requests
plotly
python-dotenv
pandas
chromadb
